build: split `Module` into a separate module
